+++ src/utils.rs
@@ -1,5 +1,5 @@
 use rand::Rng;
 
-pub fn get_rand_num(start: u8, stop: u8) -> u8 {
+pub fn get_rand_num(start: i32, stop: i32) -> i32 {
     rand::thread_rng().gen_range(start..=stop)
 }
